Create fourthFile
I have been accepting the default line of commit which is "Create fourthFile". But this time I am trying to commit and create a new branch using the available option provided at the foot of the page. And when I click this option, a new branch icon popped up with a default branch name. I am going to accept it.